+++ client/src/Pages/Dashboard/Dashboard.jsx
@@ -91,7 +91,7 @@ function Dashboard({ loading, listData, overviewLoading, overviewData, setCurren
                               <td style={{ paddingLeft: '40px' }}>{invoice?.commission} $ </td>
                               <td>{invoice?.status}</td>
                               <td>{invoice?.createdAt
-                                ? new Date(invoice?.createdAt).toLocaleDateString('en-GB', { day: 'numeric', month: 'short', year: 'numeric' })
+                                ? new Date(invoice?.createdAt).toLocaleDateString('en-GB', { day: 'numeric', month: 'long', year: 'numeric' })
                                 : 'N/A'}</td>
                             </tr>
                           ))}
